General Interface for C5.0 Rule-Based Classification Models

C5_rules() is a way to generate a specification of a model before fitting. The main arguments for the model are:

trees: The number of sequential models included in the ensemble (rules are derived from an initial set of boosted trees).

min_n: The minimum number of data points in a node that are required for the node to be split further.

These arguments are converted to their specific names at the time that the model is fit. Other options and argument can be set using parsnip::set_engine(). If left to their defaults here (NULL), the values are taken from the underlying model functions. If parameters need to be modified, update() can be used in lieu of recreating the object from scratch.

C5_rules(mode = "classification", trees = NULL, min_n = NULL)

# S3 method for C5_rules
update(
  object,
  parameters = NULL,
  trees = NULL,
  min_n = NULL,
  fresh = FALSE,
  ...
)

Arguments

mode
A single character string for the type of model. The only possible value for this model is "classification".

trees
A non-negative integer (no greater than 100 for the number of members of the ensemble.

min_n
An integer greater than one zero and nine for the minimum number of data points in a node that are required for the node to be split further.

object
A C5_rules model specification.

parameters
A 1-row tibble or named list with main parameters to update. If the individual arguments are used, these will supersede the values in parameters. Also, using engine arguments in this object will result in an error.

fresh
A logical for whether the arguments should be modified in-place or replaced wholesale.

...
Not used for update().

Details

The only availible engine is "C5.0".

Engine Details

Engines may have pre-set default arguments when executing the model fit call. For this type of model, the template of the fit calls are:

C5.0

See also

Examples

C5_rules()

#> C5.0 Model Specification (classification)
#> 
#> Computational engine: C5.0 
#> 

# Parameters can be represented by a placeholder:
C5_rules(trees = 7)

#> C5.0 Model Specification (classification)
#> 
#> Main Arguments:
#>   trees = 7
#> 
#> Computational engine: C5.0 
#> 

model <- C5_rules(trees = 10, min_n = 2)
model

#> C5.0 Model Specification (classification)
#> 
#> Main Arguments:
#>   trees = 10
#>   min_n = 2
#> 
#> Computational engine: C5.0 
#> 

update(model, trees = 1)

#> C5.0 Model Specification (classification)
#> 
#> Main Arguments:
#>   trees = 1
#>   min_n = 2
#> 
#> Computational engine: C5.0 
#> 

update(model, trees = 1, fresh = TRUE)

#> C5.0 Model Specification (classification)
#> 
#> Main Arguments:
#>   trees = 1
#> 
#> Computational engine: C5.0 
#>