General Interface for Cubist Rule-Based Regression Models

cubist_rules() is a way to generate a specification of a model before fitting. The main arguments for the model are:

committees: The number of sequential models included in the ensemble (similar to the number of trees in boosting).

neighbors: The number of neighbors in the post-model instance-based adjustment.

These arguments are converted to their specific names at the time that the model is fit. Other options and argument can be set using parsnip::set_engine(). If left to their defaults here (NULL), the values are taken from the underlying model functions. If parameters need to be modified, update() can be used in lieu of recreating the object from scratch.

cubist_rules(
  mode = "regression",
  committees = NULL,
  neighbors = NULL,
  max_rules = NULL
)

# S3 method for cubist_rules
update(
  object,
  parameters = NULL,
  committees = NULL,
  neighbors = NULL,
  max_rules = NULL,
  fresh = FALSE,
  ...
)

Arguments

mode
A single character string for the type of model. The only possible value for this model is "regression".

committees
A non-negative integer (no greater than 100 for the number of members of the ensemble.

neighbors
An integer between zero and nine for the number of training set instances that are used to adjust the model-based prediction.

max_rules
The largest number of rules.

object
A Cubist model specification.

parameters
A 1-row tibble or named list with main parameters to update. If the individual arguments are used, these will supersede the values in parameters. Also, using engine arguments in this object will result in an error.

fresh
A logical for whether the arguments should be modified in-place or replaced wholesale.

...
Not used for update().

Details

The only availible engine is "Cubist".

Engine Details

Engines may have pre-set default arguments when executing the model fit call. For this type of model, the template of the fit calls are:

Cubist

See also

Examples

cubist_rules()

#> Cubist Model Specification (regression)
#> 
#> Computational engine: Cubist 
#> 

# Parameters can be represented by a placeholder:
cubist_rules(committees = 7)

#> Cubist Model Specification (regression)
#> 
#> Main Arguments:
#>   committees = 7
#> 
#> Computational engine: Cubist 
#> 

model <- cubist_rules(committees = 10, neighbors = 2)
model

#> Cubist Model Specification (regression)
#> 
#> Main Arguments:
#>   committees = 10
#>   neighbors = 2
#> 
#> Computational engine: Cubist 
#> 

update(model, committees = 1)

#> Cubist Model Specification (regression)
#> 
#> Main Arguments:
#>   committees = 1
#>   neighbors = 2
#> 
#> Computational engine: Cubist 
#> 

update(model, committees = 1, fresh = TRUE)

#> Cubist Model Specification (regression)
#> 
#> Main Arguments:
#>   committees = 1
#> 
#> Computational engine: Cubist 
#>