RuleFit models

rule_fit() defines a model that derives simple feature rules from a tree ensemble and uses them as features to a regularized model.

The engine for this model is:

xrf (default)

More information on how parsnip is used for modeling is at https://www.tidymodels.org/.

rule_fit(
  mode = "unknown",
  mtry = NULL,
  trees = NULL,
  min_n = NULL,
  tree_depth = NULL,
  learn_rate = NULL,
  loss_reduction = NULL,
  sample_size = NULL,
  penalty = NULL,
  engine = "xrf"
)

Arguments

mode
A single character string for the type of model. Possible values for this model are "unknown", "regression", or "classification".

mtry
An number for the number (or proportion) of predictors that will be randomly sampled at each split when creating the tree models.

trees
An integer for the number of trees contained in the ensemble.

min_n
An integer for the minimum number of data points in a node that are required for the node to be split further.

tree_depth
An integer for the maximum depth of the tree (i.e. number of splits).

learn_rate
A number for the rate at which the boosting algorithm adapts from iteration-to-iteration.

loss_reduction
A number for the reduction in the loss function required to split further .

sample_size
An number for the number (or proportion) of data that is exposed to the fitting routine.

penalty
L1 regularization parameter.

engine
A single character string specifying what computational engine to use for fitting.

Details

The RuleFit model creates a regression model of rules in two stages. The first stage uses a tree-based model that is used to generate a set of rules that can be filtered, modified, and simplified. These rules are then added as predictors to a regularized generalized linear model that can also conduct feature selection during model training.

This function only defines what type of model is being fit. Once an engine is specified, the method to fit the model is also defined.

The model is not trained or fit until the fit.model_spec() function is used with the data.

References

Friedman, J. H., and Popescu, B. E. (2008). "Predictive learning via rule ensembles." The Annals of Applied Statistics, 2(3), 916-954.

https://www.tidymodels.org, Tidy Models with R

See also

Examples

show_engines("rule_fit")
#> # A tibble: 2 × 2
#>   engine mode          
#>   <chr>  <chr>         
#> 1 xrf    classification
#> 2 xrf    regression    

rule_fit()
#> RuleFit Model Specification (unknown)
#> 
#> Computational engine: xrf 
#>