C5.0 rule-based classification models

C5_rules() defines a model that derives feature rules from a tree for prediction. A single tree or boosted ensemble can be used.

The engine for this model is:

C5.0 (default)

More information on how parsnip is used for modeling is at https://www.tidymodels.org/.

C5_rules(mode = "classification", trees = NULL, min_n = NULL, engine = "C5.0")

Arguments

mode
A single character string for the type of model. The only possible value for this model is "classification".

trees
A non-negative integer (no greater than 100 for the number of members of the ensemble.

min_n
An integer greater than one zero and nine for the minimum number of data points in a node that are required for the node to be split further.

engine
A single character string specifying what computational engine to use for fitting.

Details

C5.0 is a classification model that is an extension of the C4.5 model of Quinlan (1993). It has tree- and rule-based versions that also include boosting capabilities. C5_rules() enables the version of the model that uses a series of rules (see the examples below). To make a set of rules, an initial C5.0 tree is created and flattened into rules. The rules are pruned, simplified, and ordered. Rule sets are created within each iteration of boosting.

This function only defines what type of model is being fit. Once an engine is specified, the method to fit the model is also defined.

The model is not trained or fit until the fit.model_spec() function is used with the data.

References

Quinlan R (1993). C4.5: Programs for Machine Learning. Morgan Kaufmann Publishers.

https://www.tidymodels.org, Tidy Models with R

See also

Examples

show_engines("C5_rules")
#> # A tibble: 1 × 2
#>   engine mode          
#>   <chr>  <chr>         
#> 1 C5.0   classification

C5_rules()
#> C5.0 Model Specification (classification)
#> 
#> Computational engine: C5.0 
#>