multi_predict() methods for rule-based models

multi_predict() methods for rule-based models

Usage,

# S3 method for `_C5_rules`
multi_predict(object, new_data, type = NULL, trees = NULL, ...)

# S3 method for `_cubist`
multi_predict(object, new_data, type = NULL, neighbors = NULL, ...)

# S3 method for `_xrf`
multi_predict(object, new_data, type = NULL, penalty = NULL, ...)

Arguments

object

An object of class model_fit

new_data

A rectangular data object, such as a data frame.

type

A single character value or NULL. Possible values are class" and "prob".

trees

An numeric vector of trees between one and 100.

...

Not currently used.

neighbors

An numeric vector of neighbors values between zero and nine.

penalty

Non-negative penalty values.

Value

A tibble with one row for each row of new_data. Multiple predictions are contained in a list column called .pred. That column has the standard parsnip prediction column names as well as the column with the tuning parameter values.

Details

For C5.0 rule-based models, the model fit may contain less boosting iterations than the number requested. Printing the object will show how many were used due to early stopping. This can be change using an option in C50::C5.0Control(). Beware that the number of iterations requested